elevice-backend/services/tts_service/requirements.txt
fastapi
uvicorn
google-cloud-texttospeech
python-dotenv